SBP-separation fix; change to minimize()
Fixed an error with saving SBP clauses for re-addition. The code
was not cloning the clause arrays, and those arrays were being
re-used.

Also implemented minimization-checking mod SBP.

IMPORTANT: I left two dialog box calls un-commented, because
when one of the calls is missing the final model-search loops
forever. It's probably something dumb, but it is late and I will
find the problem tomorrow.