Issue #1.
When given a trivial solution, Aluminum will now properly handle
requsts for next() minimal models (there are none). It also
produces the set of consistent facts correctly. Actually
augmenting by these facts is currently forbidden; if the user
tries to augment by a fact, they are told that any fact in the list
works without consequence, but that further exploration is disabled
on this spec.

Not the best error message. Ideally, we would tweak the lifting
process to support trivially sat. specs. But for now, this result
is at least *correct*.